requirements/base.txt
Django==1.11
django-environ==0.4.4
django-braces==1.12.0
django-crispy-forms==1.7.0
django-authtools==1.6.0
djangorestframework==3.7.7
django-allauth==0.35.0
django-bootstrap4==0.0.5
django-friendship
channels
psycopg2-binary==2.7.4
channels_redis
